+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Removed `Sieve` in favor of `QueryKit`
 * Removed `CanFilter` and `CanSort` properties. These can still be set in your code, but will no longer pollute your entities.
 * **TODO update querykit to non-prerelease**
+* Use top level routing for health checks and controllers
 
 ### Fixed